Make constraint toggling available internally, use in tabular import automatically.
